Extract local tool startup and clients into runtime
