+++ packages/agent-cli/test/Spec.hs
@@ -118,7 +118,6 @@ import qualified Agent.CLI.TUIScrollSpec as TUIScrollSpec
 import qualified Agent.CLI.TUITranscriptSpec as TUITranscriptSpec
 import qualified Agent.CLI.TUIMeasuredViewportSpec as TUIMeasuredViewportSpec
 import qualified Agent.CLI.UsageSpec as UsageSpec
-import qualified Agent.CLI.WebLspSpec as WebLspSpec
 import qualified Agent.CLI.WorktreeSpec as WorktreeSpec
 import qualified Agent.CLI.WorktreeCleanSpec as WorktreeCleanSpec
 import qualified Agent.CLI.WorktreeIgnoredSpec as WorktreeIgnoredSpec
@@ -261,7 +260,6 @@ specs = do
     TUITranscriptSpec.spec
     TUIMeasuredViewportSpec.spec
     UsageSpec.spec
-    WebLspSpec.spec
     WorktreeSpec.spec
     WorktreeCleanSpec.spec
     WorktreeIgnoredSpec.spec
