+++ packages/agent-runtime/package.nix
@@ -1,13 +1,14 @@
 { mkDerivation, aeson, agent-accounts, agent-claude
-, agent-codex-dialect, agent-connectivity, agent-core, agent-gemini
-, agent-grok-build-dialect, agent-integration-api, agent-json
-, agent-mcp, agent-openai, agent-openrouter, agent-process
-, agent-responses, agent-responses-types, agent-server-client
-, agent-store, agent-tools, agent-xai, async, base
-, base64-bytestring, bytestring, containers, crypton, directory
-, entropy, filelock, filepath, hasql-pool, hspec, http-client
-, http-client-tls, http-types, lib, memory, network, network-uri
-, process, QuickCheck, resourcet, safe-exceptions, scientific, stm
+, agent-codex-dialect, agent-computer-use, agent-connectivity
+, agent-core, agent-gemini, agent-grok-build-dialect
+, agent-integration-api, agent-json, agent-mcp, agent-openai
+, agent-openrouter, agent-process, agent-responses
+, agent-responses-types, agent-server-client, agent-store
+, agent-tools, agent-xai, async, base, base64-bytestring
+, bytestring, containers, crypton, directory, entropy, filelock
+, filepath, hasql-pool, hspec, http-client, http-client-tls
+, http-types, lib, memory, network, network-uri, process
+, QuickCheck, resourcet, safe-exceptions, scientific, stm, tagsoup
 , text, time, transformers, unix, vector, wai, warp
 }:
 mkDerivation {
@@ -17,15 +18,15 @@ mkDerivation {
   enableSeparateDataOutput = true;
   libraryHaskellDepends = [
     aeson agent-accounts agent-claude agent-codex-dialect
-    agent-connectivity agent-core agent-gemini agent-grok-build-dialect
-    agent-integration-api agent-json agent-mcp agent-openai
-    agent-openrouter agent-process agent-responses
+    agent-computer-use agent-connectivity agent-core agent-gemini
+    agent-grok-build-dialect agent-integration-api agent-json agent-mcp
+    agent-openai agent-openrouter agent-process agent-responses
     agent-responses-types agent-server-client agent-store agent-tools
     agent-xai async base base64-bytestring bytestring containers
     crypton directory entropy filelock filepath hasql-pool http-client
     http-client-tls http-types memory network network-uri process
-    resourcet safe-exceptions scientific stm text time transformers
-    unix vector wai warp
+    resourcet safe-exceptions scientific stm tagsoup text time
+    transformers unix vector wai warp
   ];
   testHaskellDepends = [
     aeson agent-accounts agent-connectivity agent-core agent-json
