+++ packages/agent-runtime/test/Spec.hs
@@ -4,6 +4,8 @@ import qualified Agent.Runtime.RequestSpec as Request
 import qualified Agent.Runtime.Tools.ResourcesSpec as ToolResources
 import qualified Agent.Runtime.Tools.DialectsSpec as ToolDialects
 import qualified Agent.Runtime.Tools.StartupSpec as ToolStartup
+import qualified Agent.Runtime.Tools.LocalStartupSpec as LocalTools
+import qualified Agent.Runtime.Tools.WebLspSpec as WebLsp
 import qualified Agent.Runtime.Mcp.StartupSpec as McpStartup
 import qualified Agent.Runtime.CollaborationSpec as Collaboration
 import qualified Agent.Runtime.ProviderRuntimeSpec as ProviderRuntime
@@ -47,6 +49,8 @@ main = hspec do
     ToolDialects.spec
     ToolResources.spec
     ToolStartup.spec
+    LocalTools.spec
+    WebLsp.spec
     McpStartup.spec
     Collaboration.spec
     StartupContext.spec
