+++ scripts/check-package-boundaries.sh
@@ -27,6 +27,13 @@ if [[ -e "$cli/src/Agent/CLI/Dialects.hs" \
 fi
 
 python3 "$root/scripts/check-package-graph.py" "$root"
+if rg --line-number '\b(newWebFetchRuntime|newLspRuntime|newComputerUseRuntime|closeWebFetchRuntime|closeLspRuntime|closeComputerUseRuntime)\b' \
+    "$cli/src/Agent/CLI/Runtime/Orchestration/Tools.hs"; then
+  fail "local-tool acquisition and cleanup must remain in agent-runtime"
+fi
+if [[ -d "$cli/src/Agent/CLI/Lsp" ]] && find "$cli/src/Agent/CLI/Lsp" -name '*.hs' -print -quit | grep -q .; then
+  fail "LSP implementation helpers must remain in agent-runtime"
+fi
 if rg --line-number '^(data SessionInitialContext|resolveSessionInitialContext[^:]|agentsDiscoverOptions[^:]|mergeSkillCatalogs[^:])' \
     "$cli/src/Agent/CLI/Resume.hs" "$cli/src/Agent/CLI/StartupContext.hs" "$cli/src/Agent/CLI/Skills.hs"; then
   fail "startup context policy and catalog assembly must remain in agent-runtime"
@@ -174,6 +181,11 @@ for registration in \
 done
 
 required_files=(
+  packages/agent-runtime/src/Agent/Runtime/Tools/LocalStartup.hs
+  packages/agent-runtime/src/Agent/Runtime/Lsp.hs
+  packages/agent-runtime/src/Agent/Runtime/WebFetch.hs
+  packages/agent-runtime/test/Agent/Runtime/Tools/LocalStartupSpec.hs
+  packages/agent-runtime/test/Agent/Runtime/Tools/WebLspSpec.hs
   packages/agent-runtime/src/Agent/Runtime/Startup/Context.hs
   packages/agent-runtime/test/Agent/Runtime/Startup/ContextSpec.hs
   packages/agent-runtime/src/Agent/Runtime/Mcp/Startup.hs
